fix(audit): 5 more P1 findings from the 2026-08-05 full audit (MODELS/PREPROCESSING)
- models._optimization_search.MBHOptimizer.submit_evaluations force-cast
  known_candidates to int (twice: the live tracker and the plotting-only
  temp array), corrupting/mis-tracking a continuous (non-integer)
  search_space -- confirmed live: 66/80 already-evaluated fractional
  candidates were silently re-suggested on a linspace search space.
  Removed both casts; candidates now keep whatever dtype search_space
  itself holds.
- models.tuning.CatboostParamsOptimizer's rule DSL referenced grow_policy/
  model_shrink_mode as required companion fields (allow_if_values_or/_and)
  but never added them to self.params, so ParameterSampler never actually
  sampled them -- check_condition always saw params.get(field) is None,
  permanently vetoing posterior_sampling=True and boosting_type=='Ordered'
  from the approved search space. Added both as samplable fields.
- models.additive_interaction_diagnostic._cv_score indexed
  y[train_idx]/y[test_idx] directly; for a pandas Series y with a
  non-default index this performs LABEL-based indexing instead of
  positional, silently misaligning train/test rows -- confirmed live: R^2
  dropped from 0.974 to -0.088 under a shuffled-index Series. Now uses
  .iloc for Series y, matching cv_splits' positional semantics.
- preprocessing.sibling_group_cold_start_fill's interpolate=True filled a
  missing sibling group by interpolating over the POSITIONAL RANK of
  groups (via reset_index(drop=True)), not the actual order_col value
  distance the docstring promises -- confirmed the exact audit repro
  (order=(0,1,100), values=(10,NaN,50)) returned the positional midpoint
  30.0 instead of the order-value-weighted 10.4. Re-indexes by the real
  order_col value and interpolates with method="index".
- preprocessing.category_support.smoothed_target_encode_column computed
  train_encoded via y_train.groupby(train_series) over the SAME rows it
  then mapped back onto -- every row's own target informed its own
  category's shrunk mean (in-sample, non-OOF target encoding), with no
  leakage warning unlike this package's other stateful-statistic
  functions. Added an oof=True (default) K-fold out-of-fold path; a row's
  own label never informs its own encoded value under the new default.
  test_encoded (already fit on the full train split, applied to genuinely
  held-out rows) is unaffected either way. oof=False preserves the legacy
  in-sample behavior for pure EDA/screening use.

Regression tests added for all five, each verified against the pre-fix
behavior (temp-revert -> confirm the real failure signature -> restore).

Part of the audits/full_audit_2026-08-05 fix wave; tracker updated.

+++ src/mlframe/models/_optimization_search.py
@@ -609,8 +609,10 @@ class MBHOptimizer:
             if (
                 self.n_noimproving_iters == 0 and self.plotting == OptimizationProgressPlotting.OnScoreImprovement
             ) or self.plotting == OptimizationProgressPlotting.Regular:
-                # For plotting we need up-to-date known_* arrays; include the in-progress batch.
-                known_cands_for_plot = np.concatenate([self.known_candidates, np.asarray(new_candidates_batch)]).astype(int)
+                # For plotting we need up-to-date known_* arrays; include the in-progress batch. No .astype(int):
+                # candidates come straight from search_space, which may be a continuous (non-integer) space --
+                # forcing int here would corrupt plotted positions for such spaces (see the fix below).
+                known_cands_for_plot = np.concatenate([self.known_candidates, np.asarray(new_candidates_batch)])
                 known_evals_for_plot = np.concatenate([self.known_evaluations, np.asarray(new_evaluations_batch)])
                 plot_search_state(
                     search_space=self.search_space,
@@ -643,5 +645,11 @@ class MBHOptimizer:
             # 15ms vs 5ms at 2000 submits) and is <0.1% of an optimization run dominated by the per-iteration GP fit, while
             # forcing every known_candidates/known_evaluations reader (~15 sites: reshape, [-n:] slicing, .tolist(), np.all) to
             # slice the filled prefix. Not worth the risk/complexity at current scales; revisit if histories reach 10^5+.
-            self.known_candidates = np.concatenate([self.known_candidates, np.asarray(new_candidates_batch)]).astype(int)
+            # No .astype(int): candidates are whatever values search_space holds (this class supports any
+            # search_space, not just integer-valued ones -- e.g. RFECV's integer feature-counts is the
+            # dominant use case, but not the only one). Force-casting to int silently truncated a continuous
+            # search_space's known_candidates, breaking the `next_candidate not in known_candidates_set`
+            # membership checks in suggest_candidate (set built from these same truncated ints), causing
+            # already-evaluated non-integer candidates to be silently re-suggested.
+            self.known_candidates = np.concatenate([self.known_candidates, np.asarray(new_candidates_batch)])
             self.known_evaluations = np.concatenate([self.known_evaluations, np.asarray(new_evaluations_batch)])
